+++ src/encord_active/app/common/components/interactive_plots.py
